Fix: API ID field 추가 및 id route type 변경
